Consolidate helper functions for models
