+++ sveltekit/src/routes/import/index.json.ts
@@ -3,6 +3,7 @@ import { Prisma } from '@prisma/client';
 import type { RequestEvent } from '@sveltejs/kit';
 
 import prisma from '$lib/helpers/prismaClient';
+import { getModelType } from '$lib/helpers/models';
 
 export const POST = async ({ request }: RequestEvent) => {
 	const canutinFile = await request.json();
@@ -272,46 +273,6 @@ const importFromCanutinFile = async (canutinFile: CanutinFile) => {
 
 		return { error: errorMessage, importedAccounts, importedAssets };
 	}
-};
-
-// Gets the Account or Asset type id from the name
-const getModelType = async (modelTypeName: string, isAccount: boolean) => {
-	const DEFAULT_TYPE = 'Other';
-	let modelTypeId: { id: number } | null = null;
-
-	const findModel = async (name: string) => {
-		const prismaQuery = {
-			where: {
-				name: {
-					contains: name
-				}
-			},
-			select: {
-				id: true
-			}
-		};
-
-		if (isAccount) {
-			return await prisma.accountType.findFirst({ ...prismaQuery });
-		} else {
-			return await prisma.assetType.findFirst({ ...prismaQuery });
-		}
-	};
-
-	modelTypeId = await findModel(modelTypeName);
-
-	if (!modelTypeId) {
-		modelTypeId = await findModel(DEFAULT_TYPE);
-
-		if (!modelTypeId)
-			throw new Error(
-				`The default ${
-					isAccount ? 'account' : 'asset'
-				} type "${DEFAULT_TYPE}" was not found. Is the database is setup correctly?`
-			);
-	}
-
-	return modelTypeId!.id;
 };
 
 // Gets the Transaction category id from the name
